+++ NEWS.md
@@ -1,6 +1,8 @@
 # agriutilities (development version)
 
   * `extract_rcov()` does not require any other parameter but the model.
+  * `ic_reml_asr()` implementation of the information criteria proposed by Verbyla (2019).
+  * `ic_reml_spt()` implementation of the information criteria proposed by Verbyla (2019) for SpATS models.
 
 # agriutilities 1.2.0